Add function to check for ignored providers
There will be services in the pipeline that won't operate on all
providers. Some will want to include only some providers while others
will want to exclude some. This new function will handle checking the
lists of providers to either include or to ignore.

When a list of providers to include is provided, the list of excluded
providers will be ignored.