Allow extended schemas to override base schema
By updating the extended schema with a base schema, the values of
duplicate keys are determined by the base schema. This is bad
inheritance. By starting with a copy of the base schema and updating it
with the extended schema, the extended schema is able to override any
part of the base schema it wants.