Make ignore_provider a coroutine
By making `ignore_provider` a coroutine, it can be used as a message
preprocessor. Doing so requires changing its signature to accept an
instance of `henson.Application` and the provider.

As the included and excluded providers are not retrieved from the app's
settings, this should eliminate the need for services to implement their
own wrapper around `ignore_provider`.