Add function to allow schema validation
We always validate the schema of incoming messages. This function will
centralize some of the code for that. In addition to validating the
incoming message, it can log the error.

`validate_schema` is intended to be wrapped by a coroutine registered as
a message preprocessor. Upon failure, it will raise an `Abort`, telling
Henson to stop processing the message.

Because `Abort` wasn't added to Henson until version 0.5, Henson is
being added as a requirement to ensure that only supported versions are
used.