Make ignore_provider a usable preprocessor
`ignore_provider` will take the entire message rather than just the
provider so that it can be used as a preprocessor. Rather than returning
a boolean, `henson.exceptions.Abort` will be raised for ignored
providers, with the incoming message be returned otherwise.  This was
the intention of the changes introduced in fba2eec.