Add coroutines to serialize/deserialize messages
We want to work with dictionaries in our services, but some consumers
and producers require bytes. The new `jsonify` and `nosjify` coroutines
will handle serializing and deserializing the messages so that we can
get messages into the correct format without having to force a format
upon the plugins (and our callbacks).