Serialize outgoing messages.
Ideally this should be handled by the services themselves, but
`prepare_message` can't easily be turned into a postprocessor. Due to
time constraints with testing the Ingestion Pipeline and getting out of
Carpathia, `send_message` and `send_error` will handle the serialization
for the time being.

This should be addressed as soon as possible.