Add takedown and delivery schemas
Takedowns require much less information than upserts, but they should
still be validated. A new ``takedown`` schema is added for that purpose.
Additionally, objectify and transform sometimes care that a delivery is
valid regardless of its type (e.g. when emitting or receiving a
message), so a new superschema called ``delivery`` is added for use in
those cases.

Logical change: previously, action validation tolerated case
differences. This is a lot of extra code for little benefit, so it's
being removed in favor of using the actions directly inside of the
schemas. If, in the future, we need to be more lenient with case
sensitivity, we can reintroduce this behavior.