Add coroutine to initialize incoming message
As part of changing the common message structure, we need a function
that will handle initializing incoming messages. This consists of
adding a new event to `events`, copying `message` to the previous event
(if one exists), and creating a `job_id` and `originated_at` if they
haven't previously been set.

Some of this functionality overlaps with `prepare_message` which should
be updated for the new common message structure as well.