Update function to initialize outgoing message
Along with removing the functionality that is now part of
`prepare_incoming_message`, `prepare_message` is being renamed to
`prepare_outgoing_message`.

Since two of its arguments are now unused, they are being removed. As
this is a backwards incompatable change, the name change can be
introduced without causing its own breaking change. Due to this, the
version number is being changed from 0.5 to 1.0.